requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudinary==1.36.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
